Further simplified writing tests using hex strings for instructions. had to move debug prints back to test_tx_parser file because printing did not work otherwise...
