+++ .vscode/settings.json
@@ -14,7 +14,14 @@
         "vector": "c",
         "__locale": "c",
         "cstdio": "c",
-        "system_error": "c"
+        "system_error": "c",
+        "__split_buffer": "c",
+        "array": "c",
+        "bitset": "c",
+        "initializer_list": "c",
+        "string_view": "c",
+        "unordered_map": "c",
+        "utility": "c"
     },
     "C_Cpp.clang_format_path": "${workspaceRoot}/.clang-format",
     "editor.formatOnSave": true,
